(PC-37106)[PRO] chore: move last IndividualOfferInformations subcomponents in appropriate folder
