+++ pro/src/pages/IndividualOffer/IndividualOfferInformations/components/UsefulInformationForm/WithdrawalReminder.tsx
